Submit function take max allowed runs into account
On D-Wave, any jobs which exceed a certain amount of time on the
annealer time out. So, for every annealing time, there is an optimal
number of annealing reads that should be send to the annealer.

The code now automatically determines the optimal number of reads, and
runs the problem as many time as required in order to gather required
number of statistics.