Sparse out solutions array
D-Wave returns the value 3 for unused qubits. This makes the vector
needlessly too large. We will like to see 1 bit storing state of each
qubit. That can be achieved by a custom MATLAB data class. For now, we
will change 3 to 0 and sparse out the array to save some space.